feat: turn yawpitchcontrol components to typescript
