+++ src/YawPitchControl/index.ts
@@ -1,5 +1,5 @@
 import YawPitchControl from "./YawPitchControl";
 
 export {
-	YawPitchControl
+  YawPitchControl
 };
